Revert "Devtooling 826 added the site id field (#1319)"
This reverts commit b2031cdd1f7bbee61c2d26fdaa9b32e6b87bb895.
