+++ genesyscloud/telephony_providers_edges_site/resource_genesyscloud_telephony_providers_edges_site_utils.go
@@ -521,7 +521,7 @@ func GenerateSiteResourceWithCustomAttrs(
 	callerId string,
 	callerName string,
 	otherAttrs ...string) string {
-	site := fmt.Sprintf(`resource "genesyscloud_telephony_providers_edges_site" "%s" {
+	return fmt.Sprintf(`resource "genesyscloud_telephony_providers_edges_site" "%s" {
 		name = "%s"
 		description = "%s"
 		location_id = %s
@@ -533,8 +533,6 @@ func GenerateSiteResourceWithCustomAttrs(
 		%s
 	}
 	`, siteRes, name, description, locationId, mediaModel, mediaRegionsUseLatencyBased, mediaRegions, callerId, callerName, strings.Join(otherAttrs, "\n"))
-
-	return site
 }
 
 func CheckForDefaultSite(siteName string) error {
